requirements.txt
Flask==2.2.3
boto3==1.26.0
python-dotenv==1.0.0